+++ apps/wearos/src/androidTest/kotlin/dev/marlonlom/mocca/wearos/ui/main/SmartwatchActivityTest.kt
@@ -12,8 +12,7 @@ import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.compose.ui.test.onNodeWithText
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import dev.marlonlom.mocca.wearos.presentation.feats.main.WearAppContent
-import dev.marlonlom.mocca.wearos.presentation.SmartwatchActivity
+import dev.marlonlom.mocca.wearos.ui.main.WearAppContent
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
