--- apps/wearos/src/main/kotlin/dev/marlonlom/mocca/wearos/presentation/SmartwatchActivity.kt
@@ -1,27 +1,0 @@
-/*
- * Copyright 2024 Marlonlom
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package dev.marlonlom.mocca.wearos.presentation
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
-import dev.marlonlom.mocca.wearos.presentation.feats.main.WearAppContent
-
-/**
- * Smartwatch activity class.
- *
- * @author marlonlom
- */
-class SmartwatchActivity : ComponentActivity() {
-  override fun onCreate(savedInstanceState: Bundle?) {
-    super.onCreate(savedInstanceState)
-    installSplashScreen()
-    setContent {
-      WearAppContent()
-    }
-  }
-}
